myapi/requirements.txt
# httpx==0.27.0
prometheus_fastapi_instrumentator<7.0.0
opentelemetry-instrumentation-sqlalchemy


pymysql
fastapi==0.111.0
uvicorn==0.29.0
prometheus-client
opentelemetry-distro
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-starlette
opentelemetry-instrumentation-logging
opentelemetry-exporter-otlp
sqlalchemy
prometheusrock

